+++ app/src/main/java/org/sopt/at/presentation/ui/signin/SignInActivity.kt
@@ -84,7 +84,7 @@ class SignInActivity : ComponentActivity() {
 
         setContent {
             ATSOPTANDROIDTheme {
-                LoginUi(
+                SignInScreen(
                     signUpId = userId,
                     signUpPw = userPw,
                     navigateToSignUp = {
@@ -94,142 +94,4 @@ class SignInActivity : ComponentActivity() {
             }
         }
     }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun LoginUi(
-    modifier: Modifier = Modifier,
-    signUpId: String = "",
-    signUpPw: String = "",
-    navigateToSignUp: () -> Unit = {}
-) {
-    var id by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var showPassword by remember { mutableStateOf(value = false) }
-    val context = LocalContext.current
-
-    // 스낵바
-    val scope = rememberCoroutineScope()
-    val snackbarHostState = remember { SnackbarHostState() }
-
-    val autoLogin = AutoLogin(context)
-
-    // 로그인 버튼 색깔
-    val isFormFilled = id.isNotBlank() && password.isNotBlank()
-    val loginButtonColor = if (isFormFilled) Color(0xFFFF143C) else Color(0xFF404040)
-    val loginTextColor = if (isFormFilled) Color.White else Color.Gray
-
-    val isValidProfile =
-        id == signUpId && password == signUpPw && signUpId.isNotBlank() && signUpPw.isNotBlank()
-
-    Scaffold(
-        snackbarHost = {
-            SnackbarHost(hostState = snackbarHostState)
-        }
-    ) { innerPadding ->
-        Column(
-            modifier = modifier
-                .fillMaxSize()
-                .background(color = Color.Black)
-                .padding(innerPadding)
-                .padding(top = 100.dp)
-                .padding(horizontal = 15.dp)
-        ) {
-            Text(
-                text = stringResource(R.string.sign_in_title),
-                color = Color.White,
-                fontSize = 20.sp,
-                fontWeight = FontWeight.Bold
-            )
-
-            // 아이디 입력 창
-            IdTextField(
-                id = id,
-                onIdChange = { id = it },
-                modifier = Modifier
-                    .padding(top = 20.dp)
-            )
-
-            // 비밀번호 입력 창
-            PasswordTextField(
-                password = password,
-                onPasswordChange = { password = it },
-                showPassword = showPassword,
-                onTogglePasswordVisibility = { showPassword = !showPassword },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 10.dp)
-            )
-
-            // 로그인하기 버튼
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 20.dp)
-                    .background(loginButtonColor, RoundedCornerShape(5.dp))
-                    .padding(vertical = 14.dp)
-                    .noRippleClickable {
-                        if (isValidProfile) {
-                            // 회원 정보가 유효 하지 않을 시 스낵바
-                            scope.launch {
-                                snackbarHostState.showSnackbar("회원 정보가 유효하지 않습니다.")
-                            }
-                        } else { // 회원가입 정보 일치 시
-                            // 자동 로그인 정보 저장
-                            autoLogin.saveLoginInfo(id, password)
-                            // MyActivity로 이동
-                            val intent = Intent(context, MyActivity::class.java).apply {
-                                putExtra("Profile", id)
-                                flags =
-                                    Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
-                            }
-                            context.startActivity(intent)
-                        }
-                    },
-                contentAlignment = Alignment.Center
-            ) {
-                Text(
-                    text = stringResource(R.string.button_sign_in),
-                    fontWeight = FontWeight.Bold,
-                    color = loginTextColor
-                )
-            }
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 30.dp)
-                    .padding(top = 35.dp)
-                    .height(IntrinsicSize.Min),
-                horizontalArrangement = Arrangement.SpaceEvenly,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = stringResource(R.string.button_find_id),
-                    color = Color.Gray
-                )
-                VerticalDivider(
-                    thickness = 1.dp,
-                    color = Color.Gray
-                )
-                Text(
-                    text = stringResource(R.string.button_find_pw),
-                    color = Color.Gray
-                )
-                VerticalDivider(
-                    thickness = 1.dp,
-                    color = Color.Gray
-                )
-                Text(
-                    text = stringResource(R.string.button_sign_up),
-                    color = Color.Gray,
-                    modifier = Modifier.clickable {
-                        navigateToSignUp()
-                    }
-                )
-            }
-        }
-    }
-
 }
